+++ exporter/opentelemetry-exporter-otlp/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Add Gzip compression for exporter
   ([#1141](https://github.com/open-telemetry/opentelemetry-python/pull/1141))
+- OTLP exporter: Handle error case when no credentials supplied
+  ([#1366](https://github.com/open-telemetry/opentelemetry-python/pull/1366))
 ## Version 0.15b0
 
 Released 2020-11-02